Incorporating custom NengoDL models and minor modifications to the plots.
We still need to optimize the code for accepting general NengoDL models. At the moment, we stick to the original VPR-Bench structure and hardcode the custom VPR techniques. In addition, we need to optimize the plots as well. Especially the plot legends must be adjusted accordingly to fit into the surrounding white space. Moreover, the fonts, the thicks, and the curves must match the LaTeX level quality.